second commit

+++ src/main/java/com/api/parkingcontrol/controllers/ParkingSpotController.java
@@ -20,8 +20,10 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -70,12 +72,36 @@ public class ParkingSpotController {
     }
     return ResponseEntity.status(HttpStatus.OK).body(parkingSpotModelOptional.get());
   }
-  
+
   @DeleteMapping("/{id}")
   public ResponseEntity<Object> deleteParkingSpot(@PathVariable(value = "id") UUID id) {
     Optional<ParkingSpotModel> parkingSpotModelOptional = parkingSpotService.findById(id);
+    if (!parkingSpotModelOptional.isPresent()) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Parking Spot not found.");
+    }
     parkingSpotService.delete(parkingSpotModelOptional.get());
     return ResponseEntity.status(HttpStatus.OK).body("The Parking Spot was sucessfully deleted.");
   }
 
+  @PutMapping("/{id}")
+  public ResponseEntity<Object> updateParkingSpot(@PathVariable(value = "id") UUID id,
+      @RequestBody @Valid ParkingSpotDto parkingSpotDto) {
+    Optional<ParkingSpotModel> parkingSpotModelOptional = parkingSpotService.findById(id);
+    if (!parkingSpotModelOptional.isPresent()) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Parking Spot not found.");
+    }
+
+    ParkingSpotModel parkingSpotModel = parkingSpotModelOptional.get();
+
+    parkingSpotModel.setParkingSpotNumber(parkingSpotDto.getParkingSpotNumber());
+    parkingSpotModel.setLicensePlateCar(parkingSpotDto.getLicensePlateCar());
+    parkingSpotModel.setModelCar(parkingSpotDto.getModelCar());
+    parkingSpotModel.setBrandCar(parkingSpotDto.getBrandCar());
+    parkingSpotModel.setColorCar(parkingSpotDto.getColorCar());
+    parkingSpotModel.setResponsibleName(parkingSpotDto.getResponsibleName());
+    parkingSpotModel.setApartment(parkingSpotDto.getApartment());
+    parkingSpotModel.setBlock(parkingSpotDto.getBlock());
+
+    return ResponseEntity.status(HttpStatus.OK).body(parkingSpotService.save(parkingSpotModel));
+  }
 }
